Refactor and add new methods on interfaces for new get all cart items flow
