+++ src/main/java/com/emazon/ms_stock/domain/api/IArticleServicePort.java
@@ -1,6 +1,7 @@
 package com.emazon.ms_stock.domain.api;
 
 import com.emazon.ms_stock.application.dto.ItemQuantityDTO;
+import com.emazon.ms_stock.application.dto.out.ArticlesPriceDTO;
 import com.emazon.ms_stock.domain.model.Article;
 
 import java.util.Set;
@@ -8,4 +9,5 @@ import java.util.Set;
 public interface IArticleServicePort extends IBasicCrudServicePort<Article> {
     void addSupply(Set<ItemQuantityDTO> dto);
     void handleCartAdditionValidations(Set<ItemQuantityDTO> dto);
+    Set<ArticlesPriceDTO> getArticlesPrice(Set<Long> articleIds);
 }
