+++ src/main/java/com/emazon/ms_stock/domain/api/IBasicCrudServicePort.java
@@ -1,7 +1,7 @@
 package com.emazon.ms_stock.domain.api;
 
-import com.emazon.ms_stock.application.dto.PageDTO;
-import com.emazon.ms_stock.application.dto.PageHandler;
+import com.emazon.ms_stock.application.dto.handlers.PageDTO;
+import com.emazon.ms_stock.application.dto.handlers.PageHandler;
 
 public interface IBasicCrudServicePort<T> {
     void save(T entity);
