+++ src/main/java/com/emazon/ms_stock/domain/spi/IBasicCrudPersistencePort.java
@@ -1,9 +1,6 @@
 package com.emazon.ms_stock.domain.spi;
 
-import com.emazon.ms_stock.application.dto.PageDTO;
-import com.emazon.ms_stock.application.dto.PageHandler;
-
 public interface IBasicCrudPersistencePort<T> {
     void save(T entity);
-    PageDTO<T> findAllPageable(PageHandler page);
+
 }
